requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9,<3.0

# HTTP requests
requests==2.32.5

# Environment and configuration
python-dotenv==1.0.0

# Data processing
pandas==2.1.4

# AI/ML
openai==1.107.1
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-core==0.3.76
langchain-community==0.3.29
langgraph==0.6.7
qdrant-client==1.15.1
ddgs==9.5.5

# Utilities
tenacity==8.2.3
packaging==23.2
schedule==1.2.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
ddgs==9.5.5